v2.01b39: Мобильное приложение v2.0.0 + улучшения
Основная программа:
- Склад вставок: fix sa_select, проверка остатка, int quantity
- Логи в МСК (datetime.now вместо utcnow)
- AGENTS.md, обновлённая INSTRUCTION.md
- Запрет выдачи при нулевом остатке

Мобильное приложение (mobile/):
- 5 экранов: логин, список, детали, завершение, настройка сервера
- AbortController для всех запросов
- Cleartext traffic для HTTP на Android
- softwareKeyboardLayoutMode adjustResize
- Обработка ошибок, безопасные даты
- Версия 2.0.0

+++ backend/src/models/insert_stock.py
@@ -22,7 +22,7 @@ class InsertTransaction(Base):
     id: Mapped[int] = mapped_column(primary_key=True)
     type: Mapped[str] = mapped_column(String(20))  # incoming, outgoing, return
     product_id: Mapped[int] = mapped_column(ForeignKey("insert_products.id"))
-    quantity: Mapped[float] = mapped_column()
+    quantity: Mapped[int] = mapped_column()
     taken_by_id: Mapped[int | None] = mapped_column(ForeignKey("users.id"), nullable=True)
     location_id: Mapped[int | None] = mapped_column(ForeignKey("asset_locations.id"), nullable=True)
     destination: Mapped[str | None] = mapped_column(String(500), nullable=True)
